initial work on expanding popup to the side
